+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 ## [Unreleased]
 
 - Update `mdbook@latest` to 0.4.23.
+- Support `mdbook` on Linux (musl).
 
 ## [1.17.0] - 2022-12-14